SoundBot INFO **CLICK HERE**
As SoundBot gets closer and closer to the final release, I will solely be working on debugging and fixes due to the fact it will be running 24/7. Once SoundBot is 24/7, this will not be updated on Github, but will be updated in notepad, etc. So the final version is really final for github. I will still have other projects here, such as, astrowoot, and restyle. (restyle not yet announced). So please, don't copy or modify bla bla (look at license for that info) and learn as much as you can from this. This really wasn't possible without WayzRG. Thank you.